fix(actionlib_trajectory): Infinite while loop
No optimization is performed in order to realize multiple setpoints when
the robot is moving. The calculate position reference method will return
whether it was able to calculate references based on the goal and the
current state. If this fails, all goals will be set to aborted since the
goal combination results to this failure, not one specific goal. If this
is not performing well in reality, we will still try to do a finite
optimization.